MainPage 중복 컨테이너 삭제
