--- frontend/src/renderer/containers/MainContainer.jsx
@@ -1,12 +1,0 @@
-import MainSlideComponent from "../components/Main/MainSlideComponent"
-import MainContentComponent from "../components/Main/MainContentComponent"
-const MainContainer = () => {
-  return (
-    <>
-        <MainSlideComponent />
-        <MainContentComponent/>
-    </>
-  );
-};
-
-export default MainContainer;
